renaming & amend proto
